Create executor namespace
